+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Fixed a bug where Matrix fields could load nested entries that belonged to other owner elements. ([#17551](https://github.com/craftcms/cms/issues/17551))
 - Fixed a potential user account enumeration bug when `preventUserEnumeration` was enabled.
 
 ## 5.8.1 - 2025-07-03